JPERF-11 Respect customer's log configuration.
We don't want to override the log configuration for local virtual users.
I'm using a side effect, but it's the only way that I found and is documented
https://logging.apache.org/log4j/2.x/manual/configuration.html.

I dropped support for
virtual-users.log file because it'll never be used and for virtual-users-out.log
and virtual-users-error.log files, because it was never working.