JPERF-661: Fix CI release task by providing authentication credentials via a custom secret variable
Turns out GITHUB_TOKEN does not have a permission to create tags in the repository, hence cannot be used for the release task. Instead, let's use a custom variable (as adviced here https://docs.github.com/en/actions/configuring-and-managing-workflows/authenticating-with-the-github_token#permissions-for-the-github_token) which will be set to a personal access token.