WIP include Jira instances
Include overall Data Center plan with hook timings.
Plans should be reusable with different ways of getting a `TcpHost`:
* new via Docker
* new via AWS
* existing machines

This commit injects an `Infrastructure` SPI to reuse a plan with
different infra, but to prove the concept, it has to fit the current
CloudFormation stack in aws-infrastructure / 2-nodes-dc*.yaml
The API smells a little, because a big one-shot CFN stack is split into
multiple fine-grained methods.

Alternative API could be: `fun plan(host: TcpHost) : JiraNodePlan`

Other matters to resolve:
* loadbalancer (plan?) API
* getting results out of DC instance and node hooks