JPERF-996: Allow minimal MySQL setup
New MySQL database can't be initialized with `--skip-grant-tables` and we need to add this flag for existing datasets
to recover from lack of GRANTs in most of those.

We don't need to update Jira's baseurl in new database, as it's set during Jira setup in web GUI.

New `MinimalMysqlDatabaseIT` adds around 2 min test time on my machine.

Mutating the code by removing `GRANT` from `MinimalMysqlDatabase` breaks the test.

New `MinimalMysqlJiraHomeTest` adds around 30 ms test time on my machine.

Currently `aws-infrastructure` module is not compatile with new API. An ability to skip Jira upgrade waiting is needed.
Draft can be found under https://github.com/atlassian/aws-infrastructure/pull/154

We are intentionally not cleaning up created temp dirs, as team believes it's unnecessary.

We are deprecating `MySqlDatabase` to avoid inconsistent naming propagation in the code. With this opportunity it's good to intoduce
`MysqlDatabase.Builder`, so that we can hide unnecessary constructors.